Fix No module named jupyter_client.kernelspecapp error
When Python 2 conda environment is created, it pulls the most recent IPython (currently 4.0.0)
The project seems to be in flux right now between IPython/Jupyter and referring to the old way of registering the Python 2 kernel generates an error and prevents the build.

Explicitly installing IPython 3 at least makes the image able to build.